backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
gunicorn==20.1.0
coreapi==2.3.3